web/requirements.txt
click==7.1.1
Flask==1.1.2
flask-sqlacodegen==1.1.7
Flask-SQLAlchemy==2.4.1
inflect==4.1.0
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
PyMySQL==0.9.3
python-dotenv==0.12.0
SQLAlchemy==1.3.16
Werkzeug==1.0.1
yapf==0.29.0
